Move theme logic from Instance to external helper
