+++ PerformanceOverlay/Controller.cs
@@ -26,7 +26,6 @@ namespace PerformanceOverlay
 
         public Controller()
         {
-            Instance.Initialize();
             Instance.OnUninstall(() =>
             {
                 startupManager.Startup = false;
@@ -92,7 +91,7 @@ namespace PerformanceOverlay
             exitItem.Click += ExitItem_Click;
 
             notifyIcon = new System.Windows.Forms.NotifyIcon(components);
-            notifyIcon.Icon = Instance.IsDarkMode() ? Resources.poll_light : Resources.poll;
+            notifyIcon.Icon = WindowsDarkMode.IsDarkModeEnabled ? Resources.poll_light : Resources.poll;
             notifyIcon.Text = TitleWithVersion;
             notifyIcon.Visible = true;
             notifyIcon.ContextMenuStrip = contextMenu;
@@ -198,7 +197,7 @@ namespace PerformanceOverlay
             try
             {
                 notifyIcon.Text = TitleWithVersion + ". RTSS Version: " + OSD.Version;
-                notifyIcon.Icon = Instance.IsDarkMode() ? Resources.poll_light : Resources.poll;
+                notifyIcon.Icon = WindowsDarkMode.IsDarkModeEnabled ? Resources.poll_light : Resources.poll;
             }
             catch
             {
