+++ PowerControl/Controller.cs
@@ -40,7 +40,6 @@ namespace PowerControl
 
         public Controller()
         {
-            Instance.Initialize();
             Instance.OnUninstall(() =>
             {
                 startupManager.Startup = false;
@@ -91,7 +90,7 @@ namespace PowerControl
             exitItem.Click += ExitItem_Click;
 
             notifyIcon = new System.Windows.Forms.NotifyIcon(components);
-            notifyIcon.Icon = Instance.IsDarkMode() ? Resources.traffic_light_outline_light : Resources.traffic_light_outline;
+            notifyIcon.Icon = WindowsDarkMode.IsDarkModeEnabled ? Resources.traffic_light_outline_light : Resources.traffic_light_outline;
             notifyIcon.Text = TitleWithVersion;
             notifyIcon.Visible = true;
             notifyIcon.ContextMenuStrip = contextMenu;
@@ -185,7 +184,7 @@ namespace PowerControl
             try
             {
                 notifyIcon.Text = TitleWithVersion + ". RTSS Version: " + OSD.Version;
-                notifyIcon.Icon = Instance.IsDarkMode() ? Resources.traffic_light_outline_light : Resources.traffic_light_outline;
+                notifyIcon.Icon = WindowsDarkMode.IsDarkModeEnabled ? Resources.traffic_light_outline_light : Resources.traffic_light_outline;
             }
             catch
             {
